Update `LocationAgent` for location Interface changed (#19)
* Update sample application
