+++ NuguInterface/Sources/Capability/Location/LocationAgentDelegate.swift
@@ -20,8 +20,13 @@
 
 import Foundation
 
-/// <#Description#>
+/// The methods that you use to receive location information from an associated `LocationAgent` object.
 public protocol LocationAgentDelegate: class {
-    /// <#Description#>
-    func locationAgentRequestContext() -> LocationContext
+    
+    /// It called when need location information for using nugu service.
+    ///
+    /// It is used as `ContextInfo`, thus must be get `LocationInfo` quickly and synchronous.
+    /// Mostly used when every a request through `ASRAgent` or `TextAgent`.
+    /// Best to use cached value or constant value when possible.
+    func locationAgentRequestLocationInfo() -> LocationInfo?
 }
