--- NuguInterface/Sources/Capability/Location/LocationContext.swift
@@ -1,61 +1,0 @@
-//
-//  LocationContext.swift
-//  NuguInterface
-//
-//  Created by yonghoonKwon on 2019/10/30.
-//  Copyright (c) 2019 SK Telecom Co., Ltd. All rights reserved.
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//
-
-import Foundation
-
-/// <#Description#>
-public struct LocationContext {
-    
-    /// <#Description#>
-    public enum State: String {
-        case available = "AVAILABLE"
-        case unavailable = "UNAVAILABLE"
-        case unknown = "UNKNOWN"
-    }
-    
-    /// <#Description#>
-    public struct Current {
-        /// <#Description#>
-        public let latitude: String
-        /// <#Description#>
-        public let longitude: String
-        
-        /// <#Description#>
-        /// - Parameter latitude: <#latitude description#>
-        /// - Parameter longitude: <#longitude description#>
-        public init(latitude: String, longitude: String) {
-            self.latitude = latitude
-            self.longitude = longitude
-        }
-    }
-    
-    /// <#Description#>
-    public let state: State
-    /// <#Description#>
-    public let current: Current?
-    
-    /// <#Description#>
-    /// - Parameter state: <#state description#>
-    /// - Parameter current: <#current description#>
-    public init(state: State, current: Current?) {
-        self.state = state
-        self.current = current
-    }
-}
