+++ SampleApp/Sources/Manager/NuguLocationManager.swift
@@ -27,32 +27,14 @@ final class NuguLocationManager: NSObject {
     
     private let locationManager = CLLocationManager()
     
-    private var requestLocationPermissionCompletion: (() -> Void)?
-    
-    /// LocationContext.Current value can not be provided in asynchronous way
-    /// So the LocationAgent expects for client to pass lastest cached value of LocationContext.Current
-    private var cachedLocationCurrent: LocationContext.Current?
-    
-    private var locationState: LocationContext.State {
-        return CLLocationManager.locationServicesEnabled() ? .available : .unavailable
-    }
-    
-    /// LocationContext.Current should be set and passed as nil when,
-    /// LocationContext.State != .available
-    /// PermissionContext.Permission.State != .granted
-    var locationContext: LocationContext {
-        let locationState = self.locationState
-        guard locationState == .available else {
-            cachedLocationCurrent = nil
-            return LocationContext(state: locationState, current: nil)
-        }
-        return LocationContext(state: locationState, current: cachedLocationCurrent)
-    }
+    /// `LocationAgent` v1.0 can not be provided in asynchronous way
+    /// So the `LocationAgent` expects for client to pass lastest cached value of `LocationInfo`
+    var cachedLocationInfo: LocationInfo?
     
     override init() {
         super.init()
         
-        // Set your own options for effective CLLocationManager usage
+        // Set your own options for effective `CLLocationManager` usage
         locationManager.desiredAccuracy = kCLLocationAccuracyBest
         locationManager.distanceFilter = 200
         locationManager.pausesLocationUpdatesAutomatically = false
@@ -65,16 +47,16 @@ final class NuguLocationManager: NSObject {
 extension NuguLocationManager {
     func startUpdatingLocation() {
         switch CLLocationManager.authorizationStatus() {
+        case .notDetermined:
+            locationManager.requestWhenInUseAuthorization()
         case .authorizedWhenInUse, .authorizedAlways:
             locationManager.stopUpdatingLocation()
             locationManager.startUpdatingLocation()
-        default: break
+        case .restricted, .denied:
+            // Do not something
+            break
+        @unknown default: break
         }
-    }
-    
-    func requestLocationPermission(completion: @escaping () -> Void) {
-        requestLocationPermissionCompletion = completion
-        locationManager.requestWhenInUseAuthorization()
     }
 }
 
@@ -82,8 +64,6 @@ extension NuguLocationManager {
 
 extension NuguLocationManager: CLLocationManagerDelegate {
     func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
-        requestLocationPermissionCompletion?()
-        requestLocationPermissionCompletion = nil
         switch status {
         case .authorizedWhenInUse, .authorizedAlways:
             startUpdatingLocation()
@@ -98,7 +78,10 @@ extension NuguLocationManager: CLLocationManagerDelegate {
         // Filter updated location with your own proper accuracy limit
         let horizontalAccuracy = location.horizontalAccuracy
         if horizontalAccuracy > 0 && horizontalAccuracy < 10000 {
-            cachedLocationCurrent = LocationContext.Current(latitude: String(location.coordinate.latitude), longitude: String(location.coordinate.longitude))
+            cachedLocationInfo = LocationInfo(
+                latitude: String(location.coordinate.latitude),
+                longitude: String(location.coordinate.longitude)
+            )
         }
     }
     
@@ -108,9 +91,9 @@ extension NuguLocationManager: CLLocationManagerDelegate {
         case .locationUnknown: break
         case .denied:
             locationManager.stopUpdatingLocation()
-            cachedLocationCurrent = nil
+            cachedLocationInfo = nil
         default:
-            cachedLocationCurrent = nil
+            cachedLocationInfo = nil
         }
     }
 }
